Add ability to shoot missiles and add enemy fleet
Player can now shoot missiles using the space bar. The fleet of enemy ships is now displayed at the top of the window. Enemy ships move side to side as the player shoots at them. When there is a collision between a missile and an enemy ship, they both disappear.