Include start screen and keep score
There is now a start screen so that the player does not start playing until they hit the space bar. The enemy ships now have a value that corresponds to how many points are received for shooting it down. The score is now displayed in the bottom left corner of the screen.