Add losing screen, multiple levels, and fix bug where one enemy ship takes all of your lives
There is now a screen when you lose all of your lives that tells the player that they've lost and shows them their score. There is also a screen that pops up and allows the player to continue on to another level if they eliminate all of the enemy ships. Each level increases the rate at which enemy ships drop from the sky. A bug where the player could lose all of their lives if a single enemy ship collided with them has been resolved.