Add instructions to start screen, reset enemy ships when player loses life, and show player exploding
The start screen now tells the user to use the arrow keys and space bar to play the game. Each level randomly generates a formation of enemy ships that gets larger as the player reaches higher levels. The ships now reset when a player loses a life to give the player time to recover. An image of an explosion replaces the image of the ship for a few seconds after the player loses a life.