Add leaderboard.txt file and leaderboard methods
leaderboard.txt will store the top 10 highest scores in the format 'ABC 123' with three initials, followed by a space, followed by the score on each line. Two functions have also been written: read_leaderboard and insert_leaderboard. read_leaderboard reads from the leaderboard.txt files and stores it as a list of tuples. insert_leaderboard attempts to enter a score into the leaderboard and returns True if the score is added to the leaderboard and False otherwise.