Fix bugs with leaderboard read and insert, write leaderboard write method
Read_leaderboard method was only reading in the first line. This has been fixed. Insert leaderboard method did not work if a new score was the highest ever or was equal to another score on the board. This has been fixed. Lastly, there is now a method to write the leaderboard back to the file.